Include api and model package in angular
