+++ modules/swagger-codegen/src/main/java/io/swagger/codegen/languages/TypeScriptAngularClientCodegen.java
@@ -2,6 +2,8 @@ package io.swagger.codegen.languages;
 
 import io.swagger.codegen.SupportingFile;
 
+import java.io.File;
+
 public class TypeScriptAngularClientCodegen extends AbstractTypeScriptClientCodegen {
 	
 	@Override
@@ -19,6 +21,8 @@ public class TypeScriptAngularClientCodegen extends AbstractTypeScriptClientCode
 	    modelTemplateFiles.put("model.mustache", ".ts");
 	    apiTemplateFiles.put("api.mustache", ".ts");
 	    templateDir = "TypeScript-Angular";
-	    supportingFiles.add(new SupportingFile("api.d.mustache", apiPackage, "api.d.ts"));
+	    apiPackage = "API.Client";
+	    modelPackage = "API.Client";
+	    supportingFiles.add(new SupportingFile("api.d.mustache", apiPackage().replace('.', File.separatorChar), "api.d.ts"));
 	}
 }
