+++ samples/client/petstore/typescript-angular/API/Client/UserApi.ts
@@ -2,10 +2,10 @@
 
 /* tslint:disable:no-unused-variable member-ordering */
 
-module  {
+module API.Client {
     'use strict';
 
-    export class PetApi {
+    export class UserApi {
         private basePath = '/v2';
 
         static $inject: string[] = ['$http'];
@@ -16,11 +16,192 @@ module  {
             }
         }
 
-        public updatePet (body?: Pet, extraHttpRequestParams?: any ) : ng.IHttpPromise<{}> {
-            var path = this.basePath + '/pet';
+        public createUser (body?: User, extraHttpRequestParams?: any ) : ng.IHttpPromise<{}> {
+            var path = this.basePath + '/user';
 
             var queryParameters: any = {};
             var headerParams: any = {};
+
+            var httpRequestParams: any = {
+                method: 'POST',
+                url: path,
+                json: true,
+                data: body,
+                
+                params: queryParameters,
+                headers: headerParams
+            };
+
+            if (extraHttpRequestParams) {
+                for (var k in extraHttpRequestParams) {
+                    if (extraHttpRequestParams.hasOwnProperty(k)) {
+                        httpRequestParams[k] = extraHttpRequestParams[k];
+                    }
+                }
+            }
+
+            return this.$http(httpRequestParams);
+        }
+
+        public createUsersWithArrayInput (body?: Array<User>, extraHttpRequestParams?: any ) : ng.IHttpPromise<{}> {
+            var path = this.basePath + '/user/createWithArray';
+
+            var queryParameters: any = {};
+            var headerParams: any = {};
+
+            var httpRequestParams: any = {
+                method: 'POST',
+                url: path,
+                json: true,
+                data: body,
+                
+                params: queryParameters,
+                headers: headerParams
+            };
+
+            if (extraHttpRequestParams) {
+                for (var k in extraHttpRequestParams) {
+                    if (extraHttpRequestParams.hasOwnProperty(k)) {
+                        httpRequestParams[k] = extraHttpRequestParams[k];
+                    }
+                }
+            }
+
+            return this.$http(httpRequestParams);
+        }
+
+        public createUsersWithListInput (body?: Array<User>, extraHttpRequestParams?: any ) : ng.IHttpPromise<{}> {
+            var path = this.basePath + '/user/createWithList';
+
+            var queryParameters: any = {};
+            var headerParams: any = {};
+
+            var httpRequestParams: any = {
+                method: 'POST',
+                url: path,
+                json: true,
+                data: body,
+                
+                params: queryParameters,
+                headers: headerParams
+            };
+
+            if (extraHttpRequestParams) {
+                for (var k in extraHttpRequestParams) {
+                    if (extraHttpRequestParams.hasOwnProperty(k)) {
+                        httpRequestParams[k] = extraHttpRequestParams[k];
+                    }
+                }
+            }
+
+            return this.$http(httpRequestParams);
+        }
+
+        public loginUser (username?: string, password?: string, extraHttpRequestParams?: any ) : ng.IHttpPromise<string> {
+            var path = this.basePath + '/user/login';
+
+            var queryParameters: any = {};
+            var headerParams: any = {};
+
+            if (username !== undefined) {
+                queryParameters['username'] = username;
+            }
+
+            if (password !== undefined) {
+                queryParameters['password'] = password;
+            }
+
+            var httpRequestParams: any = {
+                method: 'GET',
+                url: path,
+                json: true,
+                
+                params: queryParameters,
+                headers: headerParams
+            };
+
+            if (extraHttpRequestParams) {
+                for (var k in extraHttpRequestParams) {
+                    if (extraHttpRequestParams.hasOwnProperty(k)) {
+                        httpRequestParams[k] = extraHttpRequestParams[k];
+                    }
+                }
+            }
+
+            return this.$http(httpRequestParams);
+        }
+
+        public logoutUser (extraHttpRequestParams?: any ) : ng.IHttpPromise<{}> {
+            var path = this.basePath + '/user/logout';
+
+            var queryParameters: any = {};
+            var headerParams: any = {};
+
+            var httpRequestParams: any = {
+                method: 'GET',
+                url: path,
+                json: true,
+                
+                params: queryParameters,
+                headers: headerParams
+            };
+
+            if (extraHttpRequestParams) {
+                for (var k in extraHttpRequestParams) {
+                    if (extraHttpRequestParams.hasOwnProperty(k)) {
+                        httpRequestParams[k] = extraHttpRequestParams[k];
+                    }
+                }
+            }
+
+            return this.$http(httpRequestParams);
+        }
+
+        public getUserByName (username: string, extraHttpRequestParams?: any ) : ng.IHttpPromise<User> {
+            var path = this.basePath + '/user/{username}';
+
+            path = path.replace('{' + 'username' + '}', String(username));
+
+            var queryParameters: any = {};
+            var headerParams: any = {};
+
+            // verify required parameter 'username' is set
+            if (!username) {
+                throw new Error('Missing required parameter username when calling getUserByName');
+            }
+
+            var httpRequestParams: any = {
+                method: 'GET',
+                url: path,
+                json: true,
+                
+                params: queryParameters,
+                headers: headerParams
+            };
+
+            if (extraHttpRequestParams) {
+                for (var k in extraHttpRequestParams) {
+                    if (extraHttpRequestParams.hasOwnProperty(k)) {
+                        httpRequestParams[k] = extraHttpRequestParams[k];
+                    }
+                }
+            }
+
+            return this.$http(httpRequestParams);
+        }
+
+        public updateUser (username: string, body?: User, extraHttpRequestParams?: any ) : ng.IHttpPromise<{}> {
+            var path = this.basePath + '/user/{username}';
+
+            path = path.replace('{' + 'username' + '}', String(username));
+
+            var queryParameters: any = {};
+            var headerParams: any = {};
+
+            // verify required parameter 'username' is set
+            if (!username) {
+                throw new Error('Missing required parameter username when calling updateUser');
+            }
 
             var httpRequestParams: any = {
                 method: 'PUT',
@@ -43,209 +224,21 @@ module  {
             return this.$http(httpRequestParams);
         }
 
-        public addPet (body?: Pet, extraHttpRequestParams?: any ) : ng.IHttpPromise<{}> {
-            var path = this.basePath + '/pet';
+        public deleteUser (username: string, extraHttpRequestParams?: any ) : ng.IHttpPromise<{}> {
+            var path = this.basePath + '/user/{username}';
+
+            path = path.replace('{' + 'username' + '}', String(username));
 
             var queryParameters: any = {};
             var headerParams: any = {};
 
-            var httpRequestParams: any = {
-                method: 'POST',
-                url: path,
-                json: true,
-                data: body,
-                
-                params: queryParameters,
-                headers: headerParams
-            };
-
-            if (extraHttpRequestParams) {
-                for (var k in extraHttpRequestParams) {
-                    if (extraHttpRequestParams.hasOwnProperty(k)) {
-                        httpRequestParams[k] = extraHttpRequestParams[k];
-                    }
-                }
+            // verify required parameter 'username' is set
+            if (!username) {
+                throw new Error('Missing required parameter username when calling deleteUser');
             }
-
-            return this.$http(httpRequestParams);
-        }
-
-        public findPetsByStatus (status?: Array<string>, extraHttpRequestParams?: any ) : ng.IHttpPromise<Array<Pet>> {
-            var path = this.basePath + '/pet/findByStatus';
-
-            var queryParameters: any = {};
-            var headerParams: any = {};
-
-            if (status !== undefined) {
-                queryParameters['status'] = status;
-            }
-
-            var httpRequestParams: any = {
-                method: 'GET',
-                url: path,
-                json: true,
-                
-                params: queryParameters,
-                headers: headerParams
-            };
-
-            if (extraHttpRequestParams) {
-                for (var k in extraHttpRequestParams) {
-                    if (extraHttpRequestParams.hasOwnProperty(k)) {
-                        httpRequestParams[k] = extraHttpRequestParams[k];
-                    }
-                }
-            }
-
-            return this.$http(httpRequestParams);
-        }
-
-        public findPetsByTags (tags?: Array<string>, extraHttpRequestParams?: any ) : ng.IHttpPromise<Array<Pet>> {
-            var path = this.basePath + '/pet/findByTags';
-
-            var queryParameters: any = {};
-            var headerParams: any = {};
-
-            if (tags !== undefined) {
-                queryParameters['tags'] = tags;
-            }
-
-            var httpRequestParams: any = {
-                method: 'GET',
-                url: path,
-                json: true,
-                
-                params: queryParameters,
-                headers: headerParams
-            };
-
-            if (extraHttpRequestParams) {
-                for (var k in extraHttpRequestParams) {
-                    if (extraHttpRequestParams.hasOwnProperty(k)) {
-                        httpRequestParams[k] = extraHttpRequestParams[k];
-                    }
-                }
-            }
-
-            return this.$http(httpRequestParams);
-        }
-
-        public getPetById (petId: number, extraHttpRequestParams?: any ) : ng.IHttpPromise<Pet> {
-            var path = this.basePath + '/pet/{petId}';
-
-            path = path.replace('{' + 'petId' + '}', String(petId));
-
-            var queryParameters: any = {};
-            var headerParams: any = {};
-
-            // verify required parameter 'petId' is set
-            if (!petId) {
-                throw new Error('Missing required parameter petId when calling getPetById');
-            }
-
-            var httpRequestParams: any = {
-                method: 'GET',
-                url: path,
-                json: true,
-                
-                params: queryParameters,
-                headers: headerParams
-            };
-
-            if (extraHttpRequestParams) {
-                for (var k in extraHttpRequestParams) {
-                    if (extraHttpRequestParams.hasOwnProperty(k)) {
-                        httpRequestParams[k] = extraHttpRequestParams[k];
-                    }
-                }
-            }
-
-            return this.$http(httpRequestParams);
-        }
-
-        public updatePetWithForm (petId: string, name?: string, status?: string, extraHttpRequestParams?: any ) : ng.IHttpPromise<{}> {
-            var path = this.basePath + '/pet/{petId}';
-
-            path = path.replace('{' + 'petId' + '}', String(petId));
-
-            var queryParameters: any = {};
-            var headerParams: any = {};
-
-            // verify required parameter 'petId' is set
-            if (!petId) {
-                throw new Error('Missing required parameter petId when calling updatePetWithForm');
-            }
-
-            var httpRequestParams: any = {
-                method: 'POST',
-                url: path,
-                json: true,
-                
-                params: queryParameters,
-                headers: headerParams
-            };
-
-            if (extraHttpRequestParams) {
-                for (var k in extraHttpRequestParams) {
-                    if (extraHttpRequestParams.hasOwnProperty(k)) {
-                        httpRequestParams[k] = extraHttpRequestParams[k];
-                    }
-                }
-            }
-
-            return this.$http(httpRequestParams);
-        }
-
-        public deletePet (petId: number, apiKey?: string, extraHttpRequestParams?: any ) : ng.IHttpPromise<{}> {
-            var path = this.basePath + '/pet/{petId}';
-
-            path = path.replace('{' + 'petId' + '}', String(petId));
-
-            var queryParameters: any = {};
-            var headerParams: any = {};
-
-            // verify required parameter 'petId' is set
-            if (!petId) {
-                throw new Error('Missing required parameter petId when calling deletePet');
-            }
-
-            headerParams['apiKey'] = apiKey;
 
             var httpRequestParams: any = {
                 method: 'DELETE',
-                url: path,
-                json: true,
-                
-                params: queryParameters,
-                headers: headerParams
-            };
-
-            if (extraHttpRequestParams) {
-                for (var k in extraHttpRequestParams) {
-                    if (extraHttpRequestParams.hasOwnProperty(k)) {
-                        httpRequestParams[k] = extraHttpRequestParams[k];
-                    }
-                }
-            }
-
-            return this.$http(httpRequestParams);
-        }
-
-        public uploadFile (petId: number, additionalMetadata?: string, file?: file, extraHttpRequestParams?: any ) : ng.IHttpPromise<{}> {
-            var path = this.basePath + '/pet/{petId}/uploadImage';
-
-            path = path.replace('{' + 'petId' + '}', String(petId));
-
-            var queryParameters: any = {};
-            var headerParams: any = {};
-
-            // verify required parameter 'petId' is set
-            if (!petId) {
-                throw new Error('Missing required parameter petId when calling uploadFile');
-            }
-
-            var httpRequestParams: any = {
-                method: 'POST',
                 url: path,
                 json: true,
