+++ samples/client/petstore/typescript-angular/API/Client/api.d.ts
